requirements.txt
flask
flask-restx
sqlalchemy
psycopg2-binary
python-dotenv
uvicorn
asgiref
supabase
